+++ lib/src/app/build_app.dart
@@ -5,6 +5,8 @@ import 'package:change/model.dart';
 import 'package:change/src/app/change_app.dart';
 import 'package:change/src/app/console.dart';
 import 'package:intl/intl.dart';
+import 'package:marker/flavors.dart';
+import 'package:marker/marker.dart';
 
 CommandRunner<int> buildApp(Console console) {
   return CommandRunner<int>('change', 'Changelog manager.')
@@ -14,6 +16,7 @@ CommandRunner<int> buildApp(Console console) {
     ..addCommand(_Add(console, ChangeType.removal, 'removed'))
     ..addCommand(_Add(console, ChangeType.fix, 'fixed'))
     ..addCommand(_Add(console, ChangeType.security, 'security'))
+    ..addCommand(_Print(console))
     ..addCommand(_Release(console))
     ..argParser.addOption('changelog-path',
         abbr: 'p',
@@ -48,6 +51,41 @@ class _Add extends _ChangeCommand {
       return 1;
     }
     await _app().add(type, globalResults.arguments[1]);
+    return 0;
+  }
+}
+
+class _Print extends _ChangeCommand {
+  _Print(this.console);
+
+  final Console console;
+
+  @override
+  String get description => 'Prints changes for a released version';
+
+  @override
+  String get name => 'print';
+
+  @override
+  Future<int> run() async {
+    if (globalResults.arguments.isEmpty) {
+      console.error('Please specify released version to print.');
+      return 1;
+    }
+    var version = globalResults.arguments[1];
+    if (version.startsWith('-')) {
+      console.error('Please specify released version to print.');
+      return 1;
+    }
+    try {
+      final release = await _app().get(version);
+      final output = render(release.toMarkdown(), flavor: changelog);
+      // Strip first line, which is already part of the command or commit message header
+      console.log(output.substring(output.indexOf('\n') + 1));
+    } on StateError {
+      console.error('Version \'${version}\' not found!');
+      return 1;
+    }
     return 0;
   }
 }